requirements.txt
opcua>=0.98,<1.0
paho-mqtt>=1.6,<2.0
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
PyYAML>=6.0,<7.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
apscheduler>=3.10,<4.0
loguru>=0.7,<0.8
uvicorn>=0.24,<0.29
sse-starlette>=1.6,<1.8
fastapi>=0.105,<0.110
pandas>=2.1,<3.0
numpy>=1.26,<2.0
